Creating Prev and Next components
