+++ frontend/root_words_ui/components/Form.js
@@ -3,6 +3,9 @@ import {InputLabel, Button, Typography, Box, Card, CardContent, Grid, IconButton
 import RootWordsDisplay from "./RootWordsDisplay";
 import Chapter from "./Chapter";
 import Verse from "./Verse";
+import NextButton from "./NextButton"
+import PrevButton from "./PrevButton"
+import { getNumberofVerses } from '../mockAPI';
 
 
 export default function Form() { 
@@ -13,14 +16,13 @@ export default function Form() {
   const handleVerseChange = (event) => {setVerse(event.target.value)};
   React.useEffect(() => {(currentChapter === '') ? setVerse('') : handleVerseChange})
 
-
   //Find Root Button Action
   const [showRootWords, setShowRootWords] = React.useState(false)
   const onClickFindRoot = () => {setShowRootWords(true)}
 
-  //Prev & Next Button Action 
-  const getPrevVerse = () => { setVerse(currentVerse - 1)};
-  const getNextVerse = () => { setVerse(currentVerse + 1)};
+
+
+
 
   return (
   <div>
@@ -32,22 +34,18 @@ export default function Form() {
 
     </Box>
 
-    <Box display="flex" justifyContent="center" pt = {5} pb={3}>
+    <Box display="flex" justifyContent="center" pt={5} pb={3}>
     <Button type="submit" variant="contained" color="inherit" onClick={onClickFindRoot}> 
     Find Root</Button>
     </Box>
 
     <Grid container direction="row" justifyContent="center" alignItems="center" display="inline-flex">
 
-      <Grid item xs={3}><Button type="submit" variant="contained" color="inherit" onClick={getPrevVerse}> 
-      <IconButton className=" fa-solid fa-angle-left"></IconButton>Previous Verse</Button>
-      </Grid>
+      <PrevButton setVerse={setVerse} currentVerse={currentVerse}/>
 
       <Grid item xs={6} className="roots-display">{ showRootWords ? <RootWordsDisplay currentChapter={currentChapter} currentVerse={currentVerse} /> : null }</Grid> 
 
-      <Grid item xs={3}><Button type="submit" variant="contained" color="inherit" onClick={getNextVerse}> 
-      Next Verse<IconButton className=" fa-solid fa-angle-right"></IconButton></Button>
-      </Grid>
+      <NextButton setVerse={setVerse} currentVerse={currentVerse}/> 
 
     </Grid>
